+++ src/web/assets/cp/src/js/Craft.js
@@ -2626,6 +2626,14 @@ $.extend(Craft, {
   },
 
   /**
+   * Releases focus within a container.
+   * @param {Object} container
+   */
+  releaseFocusWithin: function (container) {
+    Garnish.releaseFocusWithin(container);
+  },
+
+  /**
    * Sets focus to the first focusable element within a container.
    * @param {Object} container
    */